+++ CHANGELOG.md
@@ -9,6 +9,10 @@ Fixed:
 
   * file://-URL will not be "downloaded" if already present, #165
 
+Changed:
+
+  * Logging can be configured with a file `ocrd_logging.py` in either $PWD, $HOME or /etc, #164
+
 ## [0.7.5] - 2019-08-16
 
 Fixed: